+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.1 - 2022-12-02
+### Added
+- Optional `leading` widget for `CommandPaletteAction`s which will display a Widget at the left-side of the different command palette options
+- exporting new widget 'KeyboardKeyIcon', this is the widget used to create the Keyboard Key Icons for the instructions bar and the shortcuts for each action
+- Added flag, `showInstructions`, to `CommandPaletteConfig`, which when set to true, will show the basic instructions for using the command palette, navigation, selection, and closing.
+
+### Changed
+- Flutter 3.3.7
+
 ## 0.5.0 - 2022-07-20
 ### Added
 - Open to nested action via `CommandPalette.of(context).openToAction(actionId)`